Add files via upload
 Develop a system using UML for implementing the Book Bank process. It should verify the details
of the student by the central computer. The details regarding the student will be provided to the
central computer through the administrator in the book bank and the computer will verify the
details of the student and provide approval to the office. Then the books that are needed by the
student will issue from the office to him.