Add files via upload
Develop a system using UML for implementing the Exam Registration Portal. It should verify the
details of the candidate by the central computer. The details regarding the candidate will be
provided to the central computer through the administrator and the computer will verify the details
of the candidate and provide approval. Then the hall ticket will be issued from the office to the
candidate.