Add files via upload
Develop a system using UML for implementing Stock Maintenance System. In this system, the
customer can place orders and purchase items with the aid of the stock dealer and central stock
system. These orders are verified and the items are delivered to the customer. Then the stock
details are to be updated accordingly