Add files via upload
Develop a system using UML for implementing Online Course Reservation System. This system is
organized by the central management system. The student first browses and selects the desired
course of their choice. The university then checks the availability of the seat if it is available the
student is enrolled for the course.