Add files via upload
Develop a system using UML for implementing Credit Card Processing System. In this system, the
cardholder can purchase items and pays the bill with the aid of a credit card. The cashier should
accept the card and proceed with the transaction using the central system. The bill is to be verified
and the items are delivered to the cardholder based on it.