Add files via upload
Develop a system using UML for Conference Management System. It should manage the details of
the process that will be taking place at the conference. It should work along with the organizer, who
arranges the program and central management system, which consists of all the details of the
member and the role he/she, who is going to participate in the presentation